requirements.txt
discord.py==2.6.1
requests==2.31.0